move mutation dominance headers to new subdirectory
