+++ fwdpy11/headers/fwdpy11/regions/LogNormalS.hpp
@@ -6,7 +6,6 @@
 #include <sstream>
 #include <fwdpy11/policies/mutation.hpp>
 #include "Sregion.hpp"
-#include "MutationDominance.hpp"
 
 namespace fwdpy11
 {
